+++ CHANGELOG.rst
@@ -5,9 +5,15 @@ Changelog
 0.22.0 (unreleased)
 ===================
 
+Deprecation/Removals:
+
 * Drop support for Python 2.7 and 3.6.
   Only Python>=3.6 is supported (:issue:`251`).
 * Drop support for marshmallow 3 pre-releases. Only stable versions >=2.15.1 are supported.
+* Remove ``fields.Meta``.
+
+Bug fixes:
+
 * Address ``DeprecationWarning`` raised by ``Field.fail`` on marshmallow 3.
 
 0.21.2 (2019-07-01)